+++ src/Pages/Events/Events.jsx
@@ -1,18 +1,17 @@
 import React, { useEffect } from 'react'
 import './Events.css'
-
-import im1 from './events/img (1).jpeg'
-import im2 from './events/img (2).jpeg'
-import im3 from './events/img (3).jpeg'
-import im4 from './events/img (4).jpeg'
-import im5 from './events/img (5).jpeg'
-import im6 from './events/img (6).jpeg'
-import im7 from './events/img (7).jpeg'
-import im8 from './events/img (8).jpeg'
-import im9 from './events/img (9).jpeg'
-import im10 from './events/img (10).jpeg'
-import im11 from './events/img (11).jpeg'
-import im12 from './events/img (12).jpeg'
+import im1 from '../../assets/events/img (1).jpeg'
+import im2 from '../../assets/events/img (2).jpeg'
+import im3 from '../../assets/events/img (3).jpeg'
+import im4 from '../../assets/events/img (4).jpeg'
+import im5 from '../../assets/events/img (5).jpeg'
+import im6 from '../../assets/events/img (6).jpeg'
+import im7 from '../../assets/events/img (7).jpeg'
+import im8 from '../../assets/events/img (8).jpeg'
+import im9 from '../../assets/events/img (9).jpeg'
+import im10 from '../../assets/events/img (10).jpeg'
+import im11 from '../../assets/events/img (11).jpeg'
+import im12 from '../../assets/events/img (12).jpeg'
 
 function Events() {
