+++ src/Pages/Team/team/SubTeamData.js
@@ -1,31 +1,31 @@
-import ekansh from "../../assets/subTeam/Ekansh_Tyagi.jpg"
-import kalash from "../../assets/subTeam/Kalash Pachauri.jpg"
-import vansh from '../../assets/subTeam/Vansh Mishra.jpeg'
-import vidushi from "../../assets/subTeam/Vidushi Singh.jpg"
-import suzal from "../../assets/subTeam/Suzal Chaudhary.png"
-import jesika from "../../assets/subTeam/jesika.jpeg"
-import ashutosh from "../../assets/subTeam/ashu.jpg"
-import shivangi from "../../assets/subTeam/shivangi.jpeg"
-import shubh from "../../assets/subTeam/Shubh.jpeg"
-import arpit from "../../assets/subTeam/Arpit.jpeg"
-import rohit from "../../assets/subTeam/rohitCh.jpeg"
-import harshg from "../../assets/subTeam/harshg.jpeg"
-import shruti from "../../assets/subTeam/Shruti Singh.jpeg"
-import tarun from "../../assets/subTeam/Tarun Sharma.jpeg"
-import dilkhush from "../../assets/subTeam/Dilkhush Kumar.jpg"
-import akansha from "../../assets/subTeam/Akansha Sharma.jpg"
-import diwakar from "../../assets/subTeam/Diwakar Kumar.jpg"
-import deepti from "../../assets/subTeam/Deepti Kashiv.jpg"
-import teotia from "../../assets/subTeam/Harsh Teotia.jpg"
-import ishika from "../../assets/subTeam/Ishika Bansal.jpeg"
-import priyanshu from "../../assets/subTeam/Priyanshu Gupta.jpg"
-import harshil from "../../assets/subTeam/Harshil Bhatt.jpg"
-import abhay from "../../assets/subTeam/Abhay_Tyagi.PNG"
-import tushar from "../../assets/subTeam/Tushar Yadav.jpg"
-import tushar2 from "../../assets/subTeam/Tushar Chaudhary.jpg"
-import akshat from "../../assets/subTeam/Akshat Srivastava.jpg"
-import ankit from "../../assets/subTeam/Ankit Drabu.jpg"
-import saksham from "../../assets/subTeam/SakshamD.jpeg"
+import ekansh from "../../../assets/subTeam/Ekansh_Tyagi.jpg"
+import kalash from "../../../assets/subTeam/Kalash Pachauri.jpg"
+import vansh from '../../../assets/subTeam/Vansh Mishra.jpeg'
+import vidushi from "../../../assets/subTeam/Vidushi Singh.jpg"
+import suzal from "../../../assets/subTeam/Suzal Chaudhary.png"
+import jesika from "../../../assets/subTeam/jesika.jpeg"
+import ashutosh from "../../../assets/subTeam/ashu.jpg"
+import shivangi from "../../../assets/subTeam/shivangi.jpeg"
+import shubh from "../../../assets/subTeam/Shubh.jpeg"
+import arpit from "../../../assets/subTeam/Arpit.jpeg"
+import rohit from "../../../assets/subTeam/rohitCh.jpeg"
+import harshg from "../../../assets/subTeam/harshg.jpeg"
+import shruti from "../../../assets/subTeam/Shruti Singh.jpeg"
+import tarun from "../../../assets/subTeam/Tarun Sharma.jpeg"
+import dilkhush from "../../../assets/subTeam/Dilkhush Kumar.jpg"
+import akansha from "../../../assets/subTeam/Akansha Sharma.jpg"
+import diwakar from "../../../assets/subTeam/Diwakar Kumar.jpg"
+import deepti from "../../../assets/subTeam/Deepti Kashiv.jpg"
+import teotia from "../../../assets/subTeam/Harsh Teotia.jpg"
+import ishika from "../../../assets/subTeam/Ishika Bansal.jpeg"
+import priyanshu from "../../../assets/subTeam/Priyanshu Gupta.jpg"
+import harshil from "../../../assets/subTeam/Harshil Bhatt.jpg"
+import abhay from "../../../assets/subTeam/Abhay_Tyagi.PNG"
+import tushar from "../../../assets/subTeam/Tushar Yadav.jpg"
+import tushar2 from "../../../assets/subTeam/Tushar Chaudhary.jpg"
+import akshat from "../../../assets/subTeam/Akshat Srivastava.jpg"
+import ankit from "../../../assets/subTeam/Ankit Drabu.jpg"
+import saksham from "../../../assets/subTeam/SakshamD.jpeg"
 
 const SubTeamData = [
     {
